Add day 7 part 1 and part 2
Part 2 was mostly an exercise in knowing what the heck they wanted you
to do. The final implementation isn't over complicated, but the
instructions were ANYTHING but clear. Had to look on Reddit to get
clarification, but not for a solution.

Not super happy with the implementation. It is pretty sloppy, and does
some weird stuff with `recover` and how I mess with channels. Oh well.
Had to get it done to move on.