Day 12 Part 1 and Part 2
Part 2 was pretty tricky for me. I had to peek at a hint to know how to
approach this problem, but one I learned that each dimension cycles
independently of the other dimensions, it became clear. I also knew I
had to find the least common multiple of the 3 dimensions' cycles, but
instead of figuring out how to do the math of that I found the LCM
implementation online.